Improvement on User Notification

+++ src/utilities/replaceInFile.ts
@@ -2,18 +2,20 @@ import * as vscode from 'vscode';
 
 export function replaceInFile(results: any[], choice: string, replaceText: string, file: any, dom: any) {
     const pretty = require('pretty');
-    let processResult = "Success";
+    let processResult !: string;
     const jsdom = require("jsdom");
+    let isFileChanged: boolean = false;
 
     results.forEach((result : any)=> {
         switch (choice) {
             case "Set Class":
                 try {
-                    const re = /[A-Za-z]+.*/g;
+                    const re = /^[A-Za-z]+.*/g;
                     if (replaceText.match(re) === null){
                         throw new Error("Invalid class name format");
                     }
-                    result.className = replaceText.trim();  
+                    result.className = replaceText.trim();
+                    isFileChanged = true;  
                 } catch (error: any) {
                     console.log(error);
                     processResult = error.message;
@@ -21,12 +23,13 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
                 break;
             case "Set Attribute":
                 try {
-                    const re = /[A-Za-z]+\s*=\s*[A-Za-z0-9]+/g;
+                    const re = /^[A-Za-z]+\s*=\s*[A-Za-z0-9]+$/g;
                     if (replaceText.match(re) === null){
                         throw new Error("Invalid attribute-value format");
                     }
                     const attributeValuePair = replaceText.replaceAll(/"|'/g, '').split('=');
                     result.setAttribute(attributeValuePair[0].trim().replaceAll(' ',''), attributeValuePair[1].trim().replaceAll(' ', '-'));
+                    isFileChanged = true;
                 } catch (error: any) {
                     console.log(error);
                     processResult = error.message;
@@ -35,7 +38,7 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
             case "Change Tag":
                 try {
                     const newTagName = replaceText.trim().replaceAll(' ','');
-                    const re = /[A-Za-z]+/g;
+                    const re = /^[A-Za-z]+$/g;
                     if (newTagName.match(re) === null){
                         throw new Error("Invalid tag format");
                     }
@@ -58,7 +61,8 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
                         newElem.setAttribute(name, value);
                     }
                     // Replace the source element with the new element on the page 
-                    result.parentNode.replaceChild(newElem, result); 
+                    result.parentNode.replaceChild(newElem, result);
+                    isFileChanged = true; 
                 } catch (error: any) {
                     console.log(error);
                     processResult = error.message;
@@ -66,6 +70,7 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
                 break;
             case "Remove Tag":
                 result.remove();
+                isFileChanged = true;
                 break;
             case "Remove Attribute":
                 try {
@@ -74,8 +79,8 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
                     if (replaceText.match(re) === null){
                         throw new Error("Invalid attribute name format");
                     }
-
                     result.removeAttribute(replaceText);
+                    isFileChanged = true;
                 } catch (error: any) {
                     console.log(error);
                     processResult = error.message;
@@ -85,7 +90,10 @@ export function replaceInFile(results: any[], choice: string, replaceText: strin
     });
 
     try {
-        vscode.workspace.fs.writeFile(file, new TextEncoder().encode(pretty(dom.serialize(), {ocd: true})));
+        if (isFileChanged){
+            vscode.workspace.fs.writeFile(file, new TextEncoder().encode(pretty(dom.serialize(), {ocd: true})));
+            processResult = "Success";
+        }
     } catch (error: any) {
         console.log(error);
         processResult = error.message;
